+++ src/drt_scatra/scatra_timint_meshtying_strategy_s2i_elch.cpp
@@ -785,7 +785,8 @@ void SCATRA::MortarCellCalcElchSTIThermo<distypeS, distypeM>::EvaluateConditionO
         distypeS>::template EvaluateS2ICouplingODAtIntegrationPoint<distypeM>(matelectrode,
         my::ephinp_slave_, etempnp_slave_, my::ephinp_master_, my::funct_slave_, my::funct_master_,
         my::test_lm_slave_, my::test_lm_master_, kineticmodel, numelectrons, stoichiometries, kr,
-        alphaa, alphac, timefacfac, k_ss, k_ms, SCATRA::DifferentiationType::temp);
+        alphaa, alphac, timefacfac, k_ss, k_ms,
+        static_cast<int>(SCATRA::DifferentiationType::temp));
   }  // loop over integration points
 
   return;
@@ -1122,7 +1123,7 @@ void SCATRA::MortarCellCalcSTIElch<distypeS, distypeM>::EvaluateConditionOD(
         distypeS>::template EvaluateS2ICouplingODAtIntegrationPoint<distypeM>(matelectrode,
         my::ephinp_slave_[0], eelchnp_slave_, eelchnp_master_, my::funct_slave_, my::funct_master_,
         kineticmodel, kr, alphaa, alphac, peltier, timefacfac, k_ss, k_sm,
-        SCATRA::DifferentiationType::elch);
+        static_cast<int>(SCATRA::DifferentiationType::elch));
   }  // loop over integration points
 
   return;
