+++ src/drt_ssi/ssi_monolithic_evaluate_OffDiag.cpp
@@ -206,7 +206,7 @@ void SSI::ScatraStructureOffDiagCoupling::EvaluateOffDiagBlockStructureScatraDom
   eleparams.set("action", "calc_struct_stiffscalar");
 
   // linearization of structural residuals w.r.t. elch
-  eleparams.set<int>("dtype", STR::DifferentiationType::elch);
+  eleparams.set<int>("dtype", static_cast<int>(STR::DifferentiationType::elch));
 
   // set time
   eleparams.set<double>("total time", structure_->Time());
@@ -350,7 +350,7 @@ void SSI::ScatraStructureOffDiagCoupling::EvaluateScatraStructureInterfaceSlaveS
   condparams.set<int>("action", SCATRA::bd_calc_s2icoupling_od);
 
   // linearization of boundary flux w.r.t. displacement
-  condparams.set<int>("dtype", SCATRA::DifferentiationType::disp);
+  condparams.set<int>("dtype", static_cast<int>(SCATRA::DifferentiationType::disp));
 
   // number of dofset associated with displacement-related dofs on scalar transport discretization
   condparams.set<int>("ndsdisp", 1);
